[NonCompiling]New approach, hopefully almost there
