+++ Assets/TestScript1.cs
@@ -22,16 +22,3 @@ public class TestScript1 : MonoBehaviour
         
     }
 }
-
-[Serializable]
-public class TestScriptSubClass
-{
-    [SerializeField] private string privateString;
-    public string publicString;
-
-    public TestScriptSubClass(string privateString, string publicString)
-    {
-        this.privateString = privateString;
-        this.publicString = publicString;
-    }
-}
