+++ Assets/TestScriptSubClass.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fd8e2616a59d4bac803d474c40274e31
+timeCreated: 1570302496
